send entire table as one message
previously the table was split across many zmq messages which caused significant overhead on the server. now each table is sent as one giant message, putting more burden on the clients, but greatly (we hope) reducing the burden on the server.

code was also added to the client-broker to greedily slurp messages from the server. this prevents cases when the client has been busy for a few seconds, and thus there are several heartbeat messages queued, but due to fair polling, only one is read at a time. it improves performance when the client is constantly busy.