massive re-work of bgpwatcher, store, and view
now the store and view are tightly integrated into bgpwatcher, with the exception of the two previously simulated consumers (per-as visibility and bgpview status). these have been de-coupled and moved into the top-level consumers directory.  currently they do not work and will need to be turned into sort-of 'consumer-plugins' that can be called from a single configurable bgpwatcher-consumer tool.