elems are obtained from a generator in the record
previously a list of elems was retrieved, now there is a get_next_elem method
on records which appears to function in the same way as get_next_record.

under the covers it is still building a list of elems, but it is reusing memory
and so should be slightly more efficient.

closes issue #30