various fixes/improvments to the view
most notable is that now fields MUST be explicitly activated after adding them to the view.
that is, the default state for a field is INACTIVE.

also, add a pfx_add_peer method so that adding many peers for a single prefix can be optimized.

also, remove the view_int.h, which means that now all code must use the public view api (including the view_io code).