Fix bug introduced in asn32_merge_paths
When disabling the code to generate AS path strings, I also inadvertently broke the code the merges new AS paths into old AS paths. This commit fixes that problem, and also makes things slightly more efficient as the core library now never builds the AS Path string until it is needed.