Fix bug when collector starts out of sync with others
Previously we sent a sync frame every 12 views, which in normal operation meant one every hour, at the top of the hour (since we always start corsaros from a time that is a multiple of 8), but today telxatl is misbehaving and so the first RIB is at 16:33, thus causing the first view sent to be at 16:30, and bringing all processing to a grinding halt. This patch switches syncs to be sent every 3600s, but ensures that the time of the view is at the top of the hour, skipping views until that condition is met. Tested on the misbehaving telxatl and it seems to work correctly.