Fix RT bug that inflated AS path store size
Previously the Peer ID (from the view), rather than the Peer ASN, was passed to the path store when adding a RIB path to the store. The Peer ID (an incremental 16-bit number generated by the view's peer sig map, was interpreted by the path store as an ASN, and then checked against the first segment in the AS path in order to decide whether or not to create a core path. Since this Peer ID would never ("never") match the Peer ASN, the path store would incorrectly create a non-core path. Then, when an update was received for the same path, the view would correctly interrogate the store, the store would create a core path, and there would effectively be two paths in the store that resolve to the same full AS path. The good news is that this has no effect on data integrity, it simply inflates the size of the path store (by how much I'm not sure), and perhaps slows things down.