Change API to export borrowed record pointers
This is a significant change to how the BGPStream API works. Previously we made users pass in an already-created record structure that they could re-use between calls to the library. We did this to improve performance, but it turns out that it was just a waste of time since we ended up copying internal record buffers into the user's record structure. This changes the API to instead export a borrowed pointer to a record.