Signal EOS from formats when a record is "outside" all time intervals
By "outside" this really means "above" the "highest" time interval. That is, a record is read with a timestamp that is greater than the maximum interval-end value. This change assumes that resources provide records with monotonically increasing timestamps. If this is not the case, then the appropriate format should be updated accordingly.

There are two advantages of using this heuristic:
1. Users processing data from a "stream" resource (e.g., kafka) can still use time intervals (since the previous bevhavior was to try and read "all" the records from a resource before considering it EOS).
2. Users processing only the beginning of a resource (e.g., the first second of a RIB dump) need not wait for the entire dump to be read (and discarded) before EOS is reached.