Fix prefix filter bug
The patricia tree code was doing a manual copy of the prefix when creating a new node, this meant that the new "allowed_matches" field of the prefix was not copied, and as such all pefix filters were treated as "ANY" filters (rather than MORE/EXACT/LESS). This commit changes the patricia tree to use the prefix copy method.