Fix bug in resource manager when recursive open_group is needed
Previously we were recursively calling `open_group` when an open resource was sorted into a group with some closed resources (because we need to be sure that opening those resources doesn't force us to open even more resources). This caused problems when the second `open_batch` emptied a group and reaped it, because it could pull the rug out from under the first `open_batch`. I've switched from recursively calling `open_batch` with a nested `sort_batch`, to an iterative approach where we repeatedly call `open_batch` and then `sort_batch` until `sort_batch` reports that there are no "dirty" groups (groups with a mixture of open and non-open resources) in the open batch.