Fix hex string parsing code
There was some leftover code from the earlier version of the RIS-Live JSON format where we had to do some monkeying around with the raw message string. This was causing the parser to think the message was 3 bytes longer than it actually was, which was then treated as a corrupted message.