Optimize prefix filter
The old version of bgpstream_elem_prefix_match() would insert a
temporary node; collect a result set of more and less specfic prefixes;
and delete the temporary node.

The new version uses the new bgpstream_patricia_tree_walk_up_down(),
which efficiently finds the point in the tree where the prefix _would_
be inserted (without actually inserting it), and walks parents and
children from that point, running callbacks as nodes are visited
(without the need to allocate a result set), short-circuiting the walk
early if the answer is found.