Don't use "and" to mean "or" in filter language
Like terms were always logically joined by OR, so instead of allowing
them to be given multiple times separated by the misleading word "and",
we now allow a term to be given only once but take a list of arguments.
This syntax should be compatible with a planned future syntax that will
allow arbitrary "and", "or", and "not".