Allow stream and batch resources together
This updates the data interface manager to distinguish between when the resource manager is truly empty, and when it only has stream resources. In the case that it only has stream resources, then the data interface manager will periodically issue requests to the data interface to queue more resources, allowing the data interface a chance to fetch additional batch resources.