Clean up configuration of kafka support
Previously we had multiple configure options that toggled various aspects of kafka support. This commit changes configuration so that there is only one option `--without-kafka` which disables kafka support entirely. This includes disabling the kafka DI, the kafka transport module, and modifies the broker DI so that it skips any resources that require the kafka transport.

We default to building with kafka support, but if users are missing the librdkafka dependency they will be given an error telling them to install it, or to re-run configure with the `--without-kafka` option.