Fix "infinite" batch size bug when update timestamps are not in sync
Previously, we would always extend the overlap end time for a
batch of groups to open if the current group overlapped with
the batch so far in any way.

In most cases, this made sense but if we have a collector that is
providing regular updates that are slightly offset from the time
that all other collectors are reporting, we can end up in a
situation where we cannot find a group that does not overlap
with at least one other group in our current batch.

The end result: libbgpstream decides it needs to download ALL of
the updates that it has been told about by the broker, all at once.
This can (and has) lead to problems where we simply overwhelm
libcurl by trying to open too many download handles at once and
the whole system starts to break down.

The solution proposed here: only extend the overlap boundary for
a batch if the start time for the observed group matches (or is
earlier than) the timestamp of the group that started the batch.
This means that the only situation where the overlap boundary can
be extended is if the newest group has a longer duration than any
other batch members AND the group is properly aligned with the
rest of the batch.

Worked example of the failure case (for completeness sake):

Group A: 2:00 - 2:15
Group B: 2:00 - 2:05
Group C: 2:05 - 2:10
Group D: 2:10 - 2:15
Group E: 2:01 - 2:16
Group F: 2:15 - 2:30
Group G: 2:15 - 2:20
Group H: 2:20 - 2:25
Group I: 2:25 - 2:30

Group E is the troublemaker. Without Group E, our original
approach would generate two batches of "overlapping" groups.
(A, B, C and D) would cover the 2:00 - 2:15 period.
(F, G, H and I) would cover 2:15 - 2:30.

However, Group E clearly overlaps with the (A, B, C, D) batch so
instead, our old approach would now extend the period for the
first batch to 2:00 - 2:16. Now, when we see Group F, we also
have to include it in the first batch (2:15 < 2:16, so it now
overlaps). Our batch end time is now extended further to 2:30,
and will also then include Groups G, H and I when they are seen.

Now imagine that the collector that gave us the update that
created Group E is continuing to generate updates that are offset
by 1 minute from their fellow collectors (i.e. 2:16 - 2:31).
This will extend the overlap boundary to 2:31 which is almost
certainly going to overlap again with the next batch of groups from
the other collectors...

As you can see, the overlap extension can simply continue ad
infinitum even though that is not the original intent of the
batching algorithm.

The solution in this commit would instead include Group E in the
first batch with A, B, C and D but would NOT extend the period
boundary to 2:16 (because 2:01 is after the nominal group start
time of 2:00). Thus, when Group F is observed, the first batch is
still ended and a new batch is started including just Group F.