resource manager: place limit on number of unconfirmed opened elements
When opening multiple overlapping groups, it was possible to have
in excess of 60 concurrent HTTPS handshakes being attempted to
data.ris.ripe.net. This appeared to trigger some sort of DoS protection
and all of the connections would be abruptly dropped. This would
appear to the user as a failure to open update and RIB files
when using either bgpstream or pybgpstream.

With this change, when opening a group, we force an "open wait"
call (i.e. block until the element is opened successfully) every
15 elements, which seems to be enough to allow the HTTPS
connections to succeed and bgpstream programs to behave as
expected.

This change will obviously introduce a little bit of extra delay
when opening a batch of resources, but in theory it should not be
noticeable.